requirements.txt
aiofiles==25.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
click==8.3.1
dnspython==2.8.0
fastapi==0.128.0
h11==0.16.0
idna==3.11
motor==3.7.1
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pymongo==4.16.0
python-dotenv==1.2.1
PyYAML==6.0.3
redis==7.1.0
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
